app.py
from flask import Flask, jsonify, request
from flask_cors import CORS
import cv2
import numpy as np

app = Flask(__name__)
CORS(app)  # Enable CORS for all routes

# Initialize your system components
class CalibrationManager:
    def __init__(self):
        self.is_calibrated = False

    def calibrate(self, markers):
        # Implement calibration logic
        return {"status": "success", "message": "Calibration complete"}

class DartTracker:
    def track(self, frame_data):
        # Implement dart tracking logic
        return {"x": 0, "y": 0, "confidence": 0.95}

# Initialize components
calibration_manager = CalibrationManager()
dart_tracker = DartTracker()

@app.route('/api/health', methods=['GET'])
def health_check():
    return jsonify({"status": "healthy"})

@app.route('/api/calibrate', methods=['POST'])
def calibrate():
    try:
        data = request.json
        result = calibration_manager.calibrate(data.get('markers', []))
        return jsonify(result)
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route('/api/track-dart', methods=['POST'])
def track_dart():
    try:
        data = request.json
        result = dart_tracker.track(data.get('frame'))
        return jsonify(result)
    except Exception as e:
        return jsonify({"error": str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000)