requirements.txt
lask==2.0.1
flask-cors==3.0.10
numpy==2.2.3
opencv-python-headless==4.9.0.80
python-dotenv==1.0.1
